requirements.txt
requests
anthropic
python-dotenv
google-genai
feedparser